Update Docker and Compose configurations
In particular we create separate Dockerfiles for the head and compute
nodes. We specify container, host, and network names to make things
easier to refer to. And we mount a local directory for sharing code to
the containers from the host.

TODO: Create a shared volume to put the compiled binaries in for use by
all the compute node containers.